🧪 test: move form key_event_handling test to integration tests
- Form struct made public in order to do integration test
- use premade add_item Form as test form
- re-worded comments on the Form main module
