+++ src/lib.rs
@@ -8,7 +8,7 @@ pub mod prelude {
     pub(crate) use crate::dialogue::prelude::*;
     pub use crate::errors::prelude::*;
     pub(crate) use crate::fields::prelude::*;
-    pub(crate) use crate::forms::prelude::*;
+    pub use crate::forms::prelude::*;
     pub(crate) use crate::home::prelude::*;
     pub(crate) use crate::plugin::prelude::*;
     pub(crate) use crate::styles::prelude::*;
